backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Async/Queue
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# API Documentation
drf-spectacular>=0.27.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Utils
python-decouple>=3.8
Pillow>=10.1.0
icalendar>=5.0.11

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Monitoring
sentry-sdk>=1.38.0

# Email
django-anymail>=10.2
